[Feat]: WebsocketHandler + MakeDecisionWebSocket as a service
